refac silver stream
