+++ source/core/src/core/silver/infrastructure/repositories/silver_measurements_repository.py
@@ -1,13 +1,8 @@
-from typing import Callable
-
-from geh_common.domain.types.orchestration_type import OrchestrationType as GehCommonOrchestrationType
 from pyspark.sql import DataFrame
 
 import core.silver.infrastructure.config.spark_session as spark_session
-import core.utility.shared_helpers as shared_helpers
 from core.settings import StorageAccountSettings
 from core.settings.silver_settings import SilverSettings
-from core.settings.streaming_settings import StreamingSettings
 from core.silver.domain.constants.column_names.silver_measurements_column_names import SilverMeasurementsColumnNames
 from core.silver.infrastructure.config import SilverTableNames
 from core.utility import delta_table_helper
@@ -20,7 +15,7 @@ class SilverMeasurementsRepository:
         self.data_lake_settings = StorageAccountSettings().DATALAKE_STORAGE_ACCOUNT
         self.silver_container_name = SilverSettings().silver_container_name
 
-    def read_stream(self) -> DataFrame:
+    def read(self) -> DataFrame:
         spark = spark_session.initialize_spark()
         return (
             spark.readStream.format("delta")
@@ -28,25 +23,6 @@ class SilverMeasurementsRepository:
             .option("skipChangeCommits", "true")
             .table(self.table)
         )
-
-    def write_stream(
-        self,
-        measurements: DataFrame,
-        orchestration_type: GehCommonOrchestrationType,
-        batch_operation: Callable[["DataFrame", int], None],
-    ) -> bool | None:
-        checkpoint_location = self._get_streaming_checkpoint_path(orchestration_type)
-
-        write_stream = (
-            measurements.writeStream.outputMode("append")
-            .format("delta")
-            .option("checkpointLocation", checkpoint_location)
-        )
-
-        stream_settings = StreamingSettings()
-        write_stream = stream_settings.apply_streaming_settings(write_stream)
-
-        return write_stream.foreachBatch(batch_operation).start().awaitTermination()
 
     def append_if_not_exists(self, silver_measurements: DataFrame) -> None:
         """Append to the table unless there are duplicates based on all columns except 'created'.
@@ -61,14 +37,6 @@ class SilverMeasurementsRepository:
             self.table,
             self._merge_columns(),
         )
-
-    def _get_streaming_checkpoint_path(self, orchestration_type: GehCommonOrchestrationType) -> str:
-        if orchestration_type == GehCommonOrchestrationType.SUBMITTED:
-            checkpoint_name = "submitted_transactions"
-        elif orchestration_type == GehCommonOrchestrationType.MIGRATION:
-            checkpoint_name = "migrated_transactions"
-
-        return shared_helpers.get_checkpoint_path(self.data_lake_settings, self.silver_container_name, checkpoint_name)
 
     def _merge_columns(self) -> list[str]:
         return [
